+++ framework/server/src/main/java/is/codion/framework/server/AbstractRemoteEntityConnection.java
@@ -22,7 +22,6 @@ import is.codion.common.db.database.Database;
 import is.codion.common.db.exception.DatabaseException;
 import is.codion.common.event.Event;
 import is.codion.common.observable.Observer;
-import is.codion.common.rmi.server.ClientLog;
 import is.codion.common.rmi.server.RemoteClient;
 import is.codion.common.user.User;
 import is.codion.framework.db.EntityConnection;