+++ framework/server/src/main/java/is/codion/framework/server/DefaultEntityServerAdmin.java
@@ -21,7 +21,6 @@ package is.codion.framework.server;
 import is.codion.common.db.database.Database;
 import is.codion.common.db.pool.ConnectionPoolStatistics;
 import is.codion.common.logging.LoggerProxy;
-import is.codion.common.rmi.server.ClientLog;
 import is.codion.common.rmi.server.DefaultServerAdmin;
 
 import org.slf4j.Logger;