+++ framework/server/src/main/java/is/codion/framework/server/EntityServerAdmin.java
@@ -20,7 +20,6 @@ package is.codion.framework.server;
 
 import is.codion.common.db.database.Database;
 import is.codion.common.db.pool.ConnectionPoolStatistics;
-import is.codion.common.rmi.server.ClientLog;
 import is.codion.common.rmi.server.ServerAdmin;
 
 import java.rmi.RemoteException;